Decouple database connection from the ORMTestCase
